integracion de JSON, fetch y librerias

+++ js/carrito.js
@@ -74,6 +74,24 @@ function actualizarBotonesEliminar(){
 }
 
 function eliminarDelCarrito(e){
+
+    Toastify({
+        text: "Producto eliminado",
+        duration: 2500,
+        destination: "",
+        newWindow: false,
+        close: true,
+        gravity: "top", // `top` or `bottom`
+        position: "left", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+            background: "linear-gradient(to right, #1d1d92, #808080",
+            borderRadius: "2rem",
+            border: "solid #000000",
+        },
+        onClick: function(){} // Callback after click
+      }).showToast();
+
     let idBoton = e.currentTarget.id;
     const index = productosEnCarrito.findIndex(producto => producto.id === idBoton)
     
@@ -86,9 +104,22 @@ function eliminarDelCarrito(e){
 
 botonVaciar.addEventListener("click", vaciarCarrito);
 function vaciarCarrito (){
-    productosEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-    cargarProductosCarrito();
+
+    Swal.fire({
+        title: "Deseas vaciar tu carrito?",
+        icon: "question",
+        html: `se eliminarán ${productosEnCarrito.reduce((acc, producto) => acc + producto.cantidad, 0)} productos`,
+        showCloseButton: true,
+        showCancelButton: true,
+        confirmButtonText: `Si`,
+        cancelButtonText: `No`,
+      }).then((result) => {
+        if (result.isConfirmed) {
+            productosEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito))
+            cargarProductosCarrito();
+        }
+      });
 }
 
 
@@ -98,11 +129,31 @@ function actualizarTotal(){
 
 botonComprar.addEventListener("click", comprarCarrito);
 function comprarCarrito (){
+
+    Swal.fire({
+        title: "Deseas comprar estos productos?",
+        text: "Estas seguro de no incluir ningun otro producto?",
+        icon: "info",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonText: "Deseo seguir comprando",
+        confirmButtonText: "Comprar"
+      }).then((result) => {
+        if (result.isConfirmed) {
+            carritoVacio.classList.remove("disabled");
+            carritoProductos.classList.add("disabled");
+            carritoAcciones.classList.add("disabled");
+            carritoComprado.classList.add("disabled");
+          Swal.fire({
+            title: "Felicidades!, tu pedido se ha hecho",
+            text: "Te adjuntaremos el detalle y link de pago a tu correo.",
+            icon: "success"
+          });
+        }
+      });
+
     productosEnCarrito.length = 0;
     localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
     
-    carritoVacio.classList.add("disabled");
-    carritoProductos.classList.add("disabled");
-    carritoAcciones.classList.add("disabled");
-    carritoComprado.classList.remove("disabled");
+    
 }
